# Export/Perf: use numpy in extract_primitives (2-8x speedup)