@scurest 2020-08-05
Tested exporting original.glb from #1166 with default export options. The memory use also improves a lot.

* master
    * time: 118s
    * peak mem: 7.4G
* this branch
    * time: 14s
    * peak mem: 2.5G